+++ src/components/main/Header/HamburgerMenu/MenuList.tsx
@@ -2,7 +2,6 @@ import React from "react";
 import * as S from "./style";
 import { MenuCategory } from "./interface";
 
-
 export interface Menu {
   link: string;
   text: string;
@@ -12,7 +11,9 @@ function MenuList() {
   return (
     <S.MenuTextContainer>
       {MenuCategory.map(({ link, text }: Menu) => (
-        <S.MenuText key={text}><S.MenuTextLink to={link}>{text}</S.MenuTextLink></S.MenuText>
+        <S.MenuText key={text}>
+          <S.MenuTextLink to={link}>{text}</S.MenuTextLink>
+        </S.MenuText>
       ))}
     </S.MenuTextContainer>
   );
