+++ src/components/main/Header/HeaderIcon/index.tsx
@@ -3,7 +3,7 @@ import { useLocation } from "react-router";
 import { Link } from "react-router-dom";
 import { GiHamburgerMenu } from "react-icons/gi";
 import { BiBell, BiPurchaseTagAlt } from "react-icons/bi";
-import HamburgerMenu from "./HamburgerMenu";
+import HamburgerMenu from "../HamburgerMenu";
 import * as S from "./style";
 
 function HeaderIcon(): JSX.Element {
@@ -14,27 +14,31 @@ function HeaderIcon(): JSX.Element {
   const location = useLocation();
 
   useEffect(() => {
-    setMenuOpen(false)
+    setMenuOpen(false);
   }, [location.pathname, setMenuOpen]);
 
   return (
     <div>
       <S.IconContainer>
         <Link to="/alarm">
-          <BiBell color="white"  size="22"/>
+          <BiBell color="white" size="22" />
         </Link>
         <Link to="/basket">
-          <BiPurchaseTagAlt color="white"  size="22" style={{paddingLeft: "4px"}}/>
+          <BiPurchaseTagAlt
+            color="white"
+            size="22"
+            style={{ paddingLeft: "4px" }}
+          />
         </Link>
         <GiHamburgerMenu
           color="white"
           onClick={onToggleMenu}
           cursor="pointer"
           size="22"
-          style={{paddingLeft: "4px"}}
+          style={{ paddingLeft: "4px" }}
         />
       </S.IconContainer>
-      <HamburgerMenu menuOpen={menuOpen} closeMenu={closeMenu}  />
+      <HamburgerMenu menuOpen={menuOpen} closeMenu={closeMenu} />
     </div>
   );
 }
