+++ src/components/main/Header/HeaderIcon/style.ts
@@ -1,0 +1,6 @@
+import styled from "styled-components";
+
+export const IconContainer = styled.div`
+  vertical-align: middle;
+  display: flex;
+`;
