--- src/components/main/Header/Logo.tsx
@@ -1,12 +1,0 @@
-import React from "react";
-import * as S from "./style";
-
-function Logo() :JSX.Element {
-  return (
-      <S.Logo to="/">
-        LOGO
-      </S.Logo>
-  );
-}
-
-export default Logo;
