+++ src/components/main/Header/Logo/index.tsx
@@ -1,0 +1,8 @@
+import React from "react";
+import * as S from "./style";
+
+function Index(): JSX.Element {
+  return <S.Logo to="/">LOGO</S.Logo>;
+}
+
+export default Index;
