+++ src/components/main/Header/Logo/style.ts
@@ -1,0 +1,10 @@
+import styled from "styled-components";
+import { Link } from "react-router-dom";
+
+export const Logo = styled(Link)`
+  color: white;
+  font-size: 22px;
+  font-weight: 800;
+  font-family: Gilroy;
+  text-decoration: none;
+`;
